updated changes

+++ Database/object.sol
@@ -13,7 +13,7 @@ import "./delegate.sol";
 // ----------------------------------------------------------------------------
 contract ERC20Interface {
     function transfer(address to, uint tokens) public returns (bool success);
-    function balanceOf(address _owner) public constant returns (uint256 balance);
+    function balanceOf(address _owner) public constant returns (uint balance);
 }
 
 contract Recorder {
